fix: fix profil and add category change

+++ src/Controller/Front/ItemsController.php
@@ -55,9 +55,12 @@ class ItemsController extends AbstractController
     #[Route('/{id}', name: 'app_items_show', methods: ['GET'])]
     public function show(Items $item): Response
     {
-        $user = $this->getUser();
+        $connectedUser = $this->getUser();
+        if($item->getOwner() != $connectedUser){
+            return $this->redirectToRoute('front_profil_me', [], Response::HTTP_SEE_OTHER);
+        }
+
         return $this->render('front/items/show.html.twig', [
-            'user' => $user,
             'item' => $item,
         ]);
     }
@@ -65,9 +68,14 @@ class ItemsController extends AbstractController
     #[Route('/{id}/edit', name: 'app_items_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Items $item, ItemsRepository $itemsRepository, CategoryRepository $categoryRepository): Response
     {
+        $connectedUser = $this->getUser();
+        if($item->getOwner() != $connectedUser){
+            return $this->redirectToRoute('front_profil_me', [], Response::HTTP_SEE_OTHER);
+        }
+
         $form = $this->createForm(ItemsType::class, $item);
         $form->handleRequest($request);
-        $user = $this->getUser();
+
         if ($form->isSubmitted() && $form->isValid()) {
             $itemsRepository->save($item, true);
 
@@ -75,7 +83,6 @@ class ItemsController extends AbstractController
         }
 
         return $this->renderForm('front/items/edit.html.twig', [
-            'user' => $user,
             'item' => $item,
             'form' => $form,
             'categories' => $categoryRepository->findAll()
@@ -85,6 +92,11 @@ class ItemsController extends AbstractController
     #[Route('/{id}', name: 'app_items_delete', methods: ['POST'])]
     public function delete(Request $request, Items $item, ItemsRepository $itemsRepository): Response
     {
+        $connectedUser = $this->getUser();
+        if($item->getOwner() != $connectedUser){
+            return $this->redirectToRoute('front_profil_me', [], Response::HTTP_SEE_OTHER);
+        }
+
         if ($this->isCsrfTokenValid('delete' . $item->getId(), $request->request->get('_token'))) {
             $itemsRepository->remove($item, true);
         }
